+++ CHANGELOG.md
@@ -12,6 +12,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
+## 2.1.0 - 2023.04.17
+
+### Added
+
+- Added `up()` method in `PdoHandler` to create required database table.
+
+### Changed
+
+- Renamed handlers to `LocalHandler` and `PdoHandler`.
+- Miscellaneous code cleanup.
+
 ## [2.0.0] - 2023.01.26
 
 ### Added